#217: debit every entry and Add fill from the cash the next Unit is checked against (ADR 0020, RulesVersion 1.7.0)
An entry or Add was compared with the snapshot's available cash, and no
fill ever reduced it, so a Campaign's Units each passed while together
costing several times the cash. ADR 0020: "available = basis - every
actual fill cost". Every entry and Add fill is now debited once, at its
actual cost with commission, and both checks compare against what is
left, which may be negative (ADR 0020 floors the basis, not the fills).

A snapshot replaces the basis and drops only the debits of fills at or
before its own as-of. A LEAN run sends the previous close's snapshot after
the next Session's fills, so a snapshot arriving after a fill can predate
it; replacing the debit outright on arrival is how a real run proposed a
4th Unit costing 1.7 times the cash it held. Sells are never credited by a
fill; their proceeds return through a later snapshot. A fill whose cost
leaves the float64 range fails closed.

strategy.proposal.declined advances to schema 4: AvailableCash is the
figure after fill debits and may be negative. RulesVersion 1.6.0 -> 1.7.0,
with decision corpus 1.7.0.json and an unchanged fingerprint row.

Corpus: 32 scenarios differ from 1.6.0 only by the decline schema (3 -> 4);
6 differ because their fixtures were restated for the debit (the cash-skip
ladder, the two aggregate-risk overflow audits); 7 are new.

Goldens: the main backtest golden opens with $1M and a Unit costs about
$640K, so Unit 2 is now declined for insufficient cash on each of the 11
bars that reach its rung (available 364,675 against 637,800), the Campaign
holds one Unit, and it exits 5,000 shares for 8,450.00 rather than 20,000
for 17,300.00. The declared Variant's golden opens with 3,000,000 so its
scenario still has the Add Ladder's stop raises; its first Campaign is
unchanged, and its ten later entries are declined because cmd/backtest
states no snapshot after the first Campaign's exit. Tests needing four
Units run with that cash; the equity golden restates its opening cash so
the post-entry balance keeps its one-cent window.

+++ cmd/backtest/add_affordability_test.go
@@ -38,7 +38,7 @@ func TestTheBacktestNeverBuysMoreThanItsCashAndDeclinesTheAddItCannotFund(t *tes
 				t.Fatal(err)
 			}
 			if fill.Kind == event.FillKindEntry || fill.Kind == event.FillKindAdd {
-				spent += float64(fill.Quantity)*fill.Price + fill.Commission
+				spent += float64(float64(fill.Quantity)*fill.Price) + fill.Commission
 			}
 		case event.ProposalDeclinedEventType:
 			var decline event.ProposalDeclinedPayload
